fix(network-wicked): wip
rd.net.wickedonly command line parameter:
- If not present, use legacy ifup, cmdline parse scripts and net-genrules
- If present, use just wicked

+++ modules.d/35network-wicked/wicked-config.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+type getargbool > /dev/null 2>&1 || . /lib/dracut-lib.sh
+
+if ! getargbool 0 rd.net.wickedonly; then
+    return
+fi
+
 getcmdline > /tmp/cmdline.$$.conf
 wicked show-config --ifconfig dracut:cmdline:/tmp/cmdline.$$.conf > /tmp/dracut.xml
 rm -f /tmp/cmdline.$$.conf
